requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2